decouple getCollectionMetadataAction into separate function
